resistosome_pipeline/requirements.txt
biopython==1.84
freesasa==2.2.1
numpy==1.26.3
scipy==1.15.2
pandas==2.3.3
openpyxl==3.1.5
